move MockTime into common
