+++ tests/test_controlhal.py
@@ -1,0 +1,5 @@
+from controlhal import Actuator, Sensor
+
+
+def test_sensor():
+    pass
